Update Milestone project 1.ipynb
Change in some functions that already existed in the old version.

Added the random number generator to decide which player would start.

Changed the check_position_availability to be able to catch all the possible errors on input of the position wanted.

Changed the game function to have a variable (current_player) instead of two if's to decide which turn it was.